+++ docs/source/changelog.rst
@@ -11,11 +11,11 @@ Contents:
 
    changelog
 
-Version 1.2
+Master
 ============
 * Fix missing self reference `PR #11 <https://github.com/lehinevych/MediaWikiAPI/pull/11>`_ by @danielunderwood
 * Fix dependencies installable from setup.py `PR #12 <https://github.com/lehinevych/MediaWikiAPI/pull/12>`_ by @danielunderwood
-
+* Instead throwing DisambiguationError, now we have disambiguate pages property for WikipediaPage class. `Issue #18 <https://github.com/lehinevych/MediaWikiAPI/issues/18>`_
 
 Version 1.1
 ============